+++ Assets/OxGFrame/Samples~/GSFrameDemo/Scripts/DemoSC.cs
@@ -1,0 +1,57 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using Cysharp.Threading.Tasks;
+using OxGFrame.CoreFrame.GSFrame;
+
+public class DemoSC : GSBase
+{
+    public override void OnInit()
+    {
+    }
+
+    protected override async UniTask OpenSub()
+    {
+        /**
+        * Open Sub With Async
+        */
+    }
+
+    protected override void CloseSub()
+    {
+        /**
+        * Close Sub
+        */
+    }
+
+    protected override void OnShow(object obj)
+    {
+        Debug.Log("DemoSC OnShow");
+    }
+
+    protected override void OnBind()
+    {
+        /**
+         * Do Somethings Init Once Components and Events In Here
+         */
+    }
+
+    protected override void OnUpdate(float dt)
+    {
+        /**
+         * Do Update Per FrameRate
+         */
+    }
+
+    public override void OnReceiveAndRefresh(object obj = null)
+    {
+        /**
+        * Do Update Once After Protocol Handle
+        */
+    }
+
+    protected override void OnClose()
+    {
+
+    }
+}